+++ src/Enums/CsvRowWidth.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace HiFolks\DataType\Enums;
+
+enum CsvRowWidth: string
+{
+    case STRICT = "strict";
+    case PAD = "pad";
+    case SKIP = "skip";
+}